+++ AUTimeManagement.Api/src/Management.Api/Extensions/ApiWebAppExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using AUTimeManagement.Api.Business.Logic;
 using AUTimeManagement.Api.Management.Api.Security.DAL;
 using AUTimeManagement.Api.Management.Api.Security.Model;
+using Microsoft.EntityFrameworkCore;
 
 namespace AUTimeManagement.Api.Management.Api.Extensions;
 
@@ -11,20 +12,41 @@ public static class ApiWebAppExtensions
         using var scope = app.Services.CreateScope();
         using var context = scope.ServiceProvider.GetRequiredService<SecurityDbContext>();
         using var userManager = scope.ServiceProvider.GetRequiredService<UserManager<ApplicationUser>>();
-        bool isDevelopment = app.Environment.IsDevelopment();
+        using var roleManage = scope.ServiceProvider.GetRequiredService<RoleManager<IdentityRole>>();
 
-        if (isDevelopment)
+        bool isDevelopment = app.Environment.IsDevelopment();
+        bool relational = context.Database.IsRelational();
+        if (isDevelopment && !relational)
         {
             context.Database.EnsureDeleted();
         }
 
-        context.Database.EnsureCreated();
+        if (!relational)
+        {
+            context.Database.EnsureCreated();
+        }
+        else
+        {
+            await context.Database.MigrateAsync();
+        }
+
+        if (!roleManage.Roles.Any())
+        {
+            await roleManage.CreateAsync(new IdentityRole("Admin"));
+            await roleManage.CreateAsync(new IdentityRole("User"));
+        }
 
         if (isDevelopment)
         {
-            string password = app.Configuration["rootPassword"];
-            if(await userManager.FindByEmailAsync("root@root.com") is not null)
+            string? password = app.Configuration["SecurityDbConfig:rootPassword"];
+
+            if (await userManager.FindByEmailAsync("root@root.com") is null)
             {
+                if (string.IsNullOrEmpty(password))
+                {
+                    throw new ArgumentNullException(nameof(password));
+                }
+
                 ApplicationUser user = new ApplicationUser { Email = "root@root.com", EmailConfirmed = true, UserName = "root", };
                 var result = await userManager.CreateAsync(user, password);
 
@@ -34,6 +56,36 @@ public static class ApiWebAppExtensions
                 }
             }
             Console.WriteLine("Root already exists!");
+        }
+        else
+        {
+            bool isEmpty = context.Users.Any();
+            if (isEmpty)
+            {
+                string? password = app.Configuration["SecurityDbConfig:rootPassword"];
+                if (string.IsNullOrEmpty(password))
+                {
+                    throw new ArgumentNullException(nameof(password));
+                }
+                var user = new ApplicationUser
+                {
+                    Email = "root@root.com",
+                    EmailConfirmed = true,
+                    UserName = "root",
+                };
+
+                var result = await userManager.CreateAsync(user, password);
+
+                if (result.Succeeded)
+                {
+                    Console.WriteLine("Root added!");
+                }
+            }
+        }
+        var root = await userManager.FindByNameAsync("root");
+        if (root != null)
+        {
+            await userManager.AddToRoleAsync(root, "Admin");
         }
 
         app.UseBusinessLogic(isDevelopment);
